split bitbucket into its own file
