+++ oauthenticator/__init__.py
@@ -1,1 +1,2 @@
 from .github import *
+from .bitbucket import *
